refactor(layout): collapse view-mutation triggers to one verb (Part 2)
Route every layout trigger through requestLayout(). Convert ~60 direct
layoutAllViews()/layoutDevtoolsViews() calls across 20+ files to the
debounced requestLayout(), and delete the remaining direct view
mutations the completed layout pass now covers:

- devtools-panel.ts: drop the imperative devtools-cluster hides — the
  layout pass parks them at DEVTOOLS_HIDDEN_BOUNDS on close.
- runtime-core.ts: drop the devtools host setBounds; drop the
  layoutAllViews() inside attachBrowserDevtoolsToPage's setTimeout(0)
  (the deferral stays — the attach-generation logic needs it).
- window-init.ts: drop the bootstrap addChildView/setBounds for the
  singleton views — the applyStack child-list reconcile owns them, and
  the closing markDirty('stack') + requestLayout() warms the first pass.

Move requestLayout() into layout-engine.ts so layoutAllViews and
layoutDevtoolsViews can be module-private; viewport-control.ts
re-exports requestLayout to keep the import surface stable.

Gesture-begin ordering preserved: tryEnter still runs before any layout
trigger; the debounced pass reconciles focus after interactionState is
set.

Closes dex task ujebj9br.
