+++ src/main/ipc/register-canvas-ipc.ts
@@ -9,7 +9,6 @@ import { setHoverEntity, setHoveredPage } from '../runtime/runtime-core'
 import { activeTool as uiActiveTool } from '../ui-state'
 import {
   canvasOrigin,
-  layoutAllViews,
   pan,
   requestLayout,
   zoom,
@@ -175,7 +174,7 @@ export function registerCanvasIpc(): void {
 
   ipcMain.on('canvas-select-entities', (_event, entityIds: string[]) => {
     setSelectedEntities(entityIds)
-    layoutAllViews()
+    requestLayout()
   })
 
   const VALID_ENTITY_KINDS: ReadonlySet<CanvasEntityKind> = new Set<CanvasEntityKind>(
@@ -296,7 +295,7 @@ export function registerCanvasIpc(): void {
       if (!page) return
       page.metadata = setPageBrowserSizeMode(page.metadata, mode)
       scheduleWorkspaceAutosave()
-      layoutAllViews()
+      requestLayout()
     },
   )
 
@@ -418,13 +417,13 @@ export function registerCanvasIpc(): void {
       })
       commitActive()
       setHoverEntity(null)
-      layoutAllViews()
+      requestLayout()
     },
   )
 
   ipcMain.on('canvas-delete-edge', (_event, { edgeId }: { edgeId: string }) => {
     deleteEdges({ edgeIds: [edgeId] })
-    layoutAllViews()
+    requestLayout()
   })
 
   ipcMain.on('canvas-select-edge', (_event, { edgeId }: { edgeId: string | null }) => {
@@ -433,7 +432,7 @@ export function registerCanvasIpc(): void {
       return
     }
     selectEntity(edgeId, 'edge')
-    layoutAllViews()
+    requestLayout()
   })
 
   // --- File drop ---
