+++ src/main/runtime/devtools-panel.ts
@@ -4,9 +4,6 @@
 
 import { markDirty } from './layout-dirty'
 import {
-  devtoolsBackgroundView,
-  devtoolsHeaderView,
-  devtoolsResizeHandleView,
   toolbarView,
   win,
   setDevtoolsView,
@@ -25,7 +22,6 @@ import {
   setLeftSidebarOpen as setUiLeftSidebarOpen,
   setDevtoolsPanelTab as setUiDevtoolsPanelTab,
 } from '../ui-state'
-import { layoutAllViews, layoutDevtoolsViews } from './layout-engine'
 import { requestLayout } from './viewport-control'
 import { syncInspectionState } from './inspect-session'
 import { devtoolsPanelDebug } from './runtime-constants'
@@ -41,7 +37,6 @@ export function toggleLeftSidebar(): void {
   setUiLeftSidebarOpen(!uiLeftSidebarOpen())
   markDirty('sidebar', 'canvas', 'floating-ui')
   notifyDevtoolsChanged()
-  layoutAllViews()
   markDirty('stack'); requestLayout()
 }
 
@@ -56,24 +51,14 @@ export function closeDevTools(): void {
       // Ignore close races during shutdown or retargeting.
     }
   }
-  // Per-page devtools host views are hidden by the layout pass once
-  // devtools is closed — no imperative hiding needed here.
-
-  if (devtoolsBackgroundView) {
-    devtoolsBackgroundView.setBounds({ x: 0, y: 0, width: 0, height: 0 })
-  }
-  if (devtoolsHeaderView) {
-    devtoolsHeaderView.setBounds({ x: 0, y: 0, width: 0, height: 0 })
-  }
-  if (devtoolsResizeHandleView) {
-    devtoolsResizeHandleView.setBounds({ x: 0, y: 0, width: 0, height: 0 })
-  }
+  // The devtools cluster views (background / header / resize handle) and
+  // every page's devtools host view are parked off-screen by the layout
+  // pass once devtools is closed — no imperative hiding needed here.
   setDevtoolsView(null)
 
   setUiDevtoolsOpen(false)
   syncInspectionState()
   notifyDevtoolsChanged()
-  layoutDevtoolsViews()
   markDirty('stack'); requestLayout()
 }
 
@@ -93,7 +78,6 @@ export function toggleDevTools(): void {
   setUiDevtoolsOpen(true)
   notifyDevtoolsChanged()
   syncInspectionState()
-  layoutDevtoolsViews()
   markDirty('stack'); requestLayout()
   devtoolsPanelDebug('toggle:open-complete', { durationMs: Date.now() - start })
 }
@@ -116,7 +100,6 @@ export function dismissBrowserDevTools(): void {
   setUiDevtoolsPanelTab('comments')
   notifyDevtoolsChanged()
   syncInspectionState()
-  layoutDevtoolsViews()
   markDirty('stack'); requestLayout()
 }
 
@@ -130,7 +113,6 @@ export function openDevToolsForSelectedPage(): void {
   setUiDevtoolsOpen(true)
   notifyDevtoolsChanged()
   syncInspectionState()
-  layoutDevtoolsViews()
   markDirty('stack'); requestLayout()
   attachBrowserDevtoolsToPage(selectedPageIdx)
 }
@@ -145,7 +127,6 @@ export function openInspectPanel(): void {
   focusUiAnnotation(null)
   markDirty('toolbar', 'canvas', 'floating-ui')
   syncInspectionState()
-  layoutDevtoolsViews()
   markDirty('stack'); requestLayout()
 }
 
@@ -159,7 +140,6 @@ export function openCommentsPanel(annotationId?: string): void {
   focusUiAnnotation(annotationId ?? null)
   markDirty('toolbar', 'canvas', 'floating-ui')
   syncInspectionState()
-  layoutDevtoolsViews()
   markDirty('stack'); requestLayout()
 }
